+++ src/test_normal.c
@@ -40,7 +40,7 @@ void	test2()
 Given s ← sphere()\n\
 When n ← normal_at(s, point(√3/3, √3/3, √3/3))\n\
 Then n = vector(√3/3, √3/3, √3/3)\n");
-	normal = normal_at(sphere, point(sqrtf(3)/3, sqrtf(3)/3, sqrtf(3)/3));
+	normal = normal_at(sphere, point(sqrt(3)/3, sqrt(3)/3, sqrt(3)/3));
 	print_tuple(normal);
 	free(sphere);
 }
@@ -84,7 +84,7 @@ Then n = vector(0, 0.97014, -0.24254)\n");
 	sphere = sphere_new_args(point(0, 0, 0), 1, color255(255, 0, 0));
 	m = matrix4_multiply(scaling_matrix4(1, 0.5, 1), rotation_z(M_PI / 5));
 	set_transform(sphere, m);
-	normal = normal_at(sphere, point(0, sqrtf(2) / 2, -sqrtf(2) / 2));
+	normal = normal_at(sphere, point(0, sqrt(2) / 2, -sqrt(2) / 2));
 	print_tuple(normal);
 	free(sphere);
 }
@@ -120,7 +120,7 @@ And n ← vector(√2/2, √2/2, 0)\n\
 When r ← reflect(v, n)\n\
 Then r = vector(1, 0, 0)\n");
 	v = vector(0, -1, 0);
-	n = vector(sqrtf(2) / 2, sqrtf(2) / 2, 0);
+	n = vector(sqrt(2) / 2, sqrt(2) / 2, 0);
 	r = reflect(v, n);
 	print_tuple(r);
 }
